Give relation types own file
