+++ p2panda-rs/src/operation/mod.rs
@@ -10,11 +10,13 @@ mod operation;
 mod operation_encoded;
 mod operation_fields;
 mod operation_meta;
+mod relation;
 
 pub use error::{
     OperationEncodedError, OperationError, OperationFieldsError, OperationWithMetaError,
 };
 pub use operation::{AsOperation, Operation, OperationAction, OperationVersion};
 pub use operation_encoded::OperationEncoded;
-pub use operation_fields::{OperationFields, OperationValue, Relation};
+pub use operation_fields::{OperationFields, OperationValue};
 pub use operation_meta::OperationWithMeta;
+pub use relation::{Relation, RelationList};
